zone exp riconosciute ma con problema

+++ main/app_mqtt.h
@@ -11,7 +11,7 @@ esp_err_t mqtt_start(void);
 esp_err_t mqtt_stop(void);
 esp_err_t mqtt_reload_config(void);
 esp_err_t mqtt_publish_state(void);
-esp_err_t mqtt_publish_zones(uint16_t mask);
+esp_err_t mqtt_publish_zones(uint32_t mask);
 esp_err_t mqtt_publish_scenes(void);
 
 #ifdef __cplusplus
